requirements.txt
streamlit == 1.28.1
pandas == 2.1.2
numpy == 1.26.1
Pillow == 10.1.0